+++ src/components/MCQ/DropDownbox.vue
@@ -14,7 +14,7 @@
 </template>
 <script setup lang="ts">
 import { ref } from "vue";
-import { useQuizStore } from "@/plugins/CruciblePlugin/store/QuizStore";
+import { useQuizStore } from "@/store/QuizStore";
 
 // Define the props expected from the parent component
 defineProps<{
