+++ src/components/MCQ/FilterCheckbox.vue
@@ -25,8 +25,8 @@
 </template>
 
 <script setup lang="ts">
-import { SelectedTags } from "@/plugins/CruciblePlugin/types/MCQ";
-import { useQuizStore } from "@/plugins/CruciblePlugin/store/QuizStore";
+import { SelectedTags } from "@/types/MCQ";
+import { useQuizStore } from "@/store/QuizStore";
 import { getQuestionsBasedOnEnv } from "../DataAccessLayer";
 import { filterQuestionsByTags } from "../QuestionStore";
 import { computed } from "vue";
