+++ src/components/MCQ/MCQInfoPanel.vue
