+++ src/components/MCQ/MCQTimedQuiz.vue
