+++ src/main.ts
@@ -1,7 +1,7 @@
 import { createApp } from "vue";
 import { createPinia } from "pinia";
 import "./style.css";
-import CrucibleComponent from "./plugins/CruciblePlugin/components/CrucibleComponent.vue";
+import CrucibleComponent from "./components/CrucibleComponent.vue";
 
 const pinia = createPinia();
 const app = createApp(CrucibleComponent);
