+++ src/types/index.d.ts
@@ -1,0 +1,11 @@
+import "@vue/runtime-core";
+
+export type ViewerPluginOptions = {
+  dataLink?: string;
+};
+
+declare module "@vue/runtime-core" {
+  interface ComponentCustomProperties {
+    $dataLink: string;
+  }
+}
