+++ tests/components/MCQQuiz.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from "vitest";
 import { createPinia, setActivePinia } from "pinia";
-import { useQuizStore } from "@/plugins/CruciblePlugin/store/QuizStore";
+import { useQuizStore } from "../../src/store/QuizStore";
 import { questionsData as questions } from "../testSeeds";
-import MCQQuiz from "@/plugins/CruciblePlugin/components/MCQ/MCQQuiz.vue";
+import MCQQuiz from "@components/MCQ/MCQQuiz.vue";
 import { mount, VueWrapper, DOMWrapper } from "@vue/test-utils";
 import { getOptions } from "./MCQQuestion.test";
