+++ tests/components/MCQTimedQuiz.test.ts
@@ -2,8 +2,8 @@ import { VueWrapper, mount } from "@vue/test-utils";
 import { createPinia, setActivePinia } from "pinia";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { questionsData as questions } from "../testSeeds";
-import { useQuizStore } from "@/plugins/CruciblePlugin/store/QuizStore";
-import MCQTimedQuiz from "@/plugins/CruciblePlugin/components/MCQ/MCQTimedQuiz.vue";
+import { useQuizStore } from "../../src/store/QuizStore";
+import MCQTimedQuiz from "@components/MCQ/MCQTimedQuiz.vue";
 
 let wrapper: VueWrapper;
