+++ tests/components/StartPage.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach } from "vitest";
 import { mount } from "@vue/test-utils";
 import { questionsData as questions } from "../testSeeds";
-import StartPage from "@/plugins/CruciblePlugin/components/StartPage.vue";
+import StartPage from "@components/StartPage.vue";
 import { createPinia, setActivePinia } from "pinia";
 
 beforeEach(() => {
